# meeting_room_booking_system_backend/openapi.yaml
openapi: 3.0.0
paths:
  /user/check-token-expiration:
    get:
      description: 验证token是否有效
      operationId: check-token-expiration
      parameters:
        - name: token
          required: true
          in: query
          description: token
          schema:
            type: string
      responses:
        '200':
          description: 验证通过
          content:
            application/json:
              schema:
                type: string
        '500':
          description: 验证不通过
          content:
            application/json:
              schema:
                type: string
      summary: ''
      tags:
        - auth
  /user/register-captcha:
    get:
      description: 获取注册验证码
      operationId: register-captcha
      parameters:
        - name: address
          required: true
          in: query
          description: 邮箱地址
          schema:
            example: xxx@xx.com
            type: string
      responses:
        '200':
          description: 发送成功
          content:
            application/json:
              schema:
                type: string
        '400':
          description: 邮箱已存在
          content:
            application/json:
              schema:
                type: string
      summary: ''
      tags:
        - captcha
  /user/forgot-password-captcha:
    get:
      description: 获取找回密码验证码
      operationId: fotgot-captcha
      parameters:
        - name: address
          required: true
          in: query
          description: 邮箱地址
          schema:
            example: xxx@xx.com
            type: string
      responses:
        '200':
          description: 发送成功
          content:
            application/json:
              schema:
                type: string
      summary: ''
      tags:
        - captcha
  /user/update-password/captcha:
    get:
      description: 获取更改密码验证码
      operationId: update-password-captcha
      parameters: []
      responses:
        '200':
          description: 发送成功
          content:
            application/json:
              schema:
                type: string
      security:
        - bearer: []
      summary: ''
      tags:
        - captcha
  /user/update/captcha:
    get:
      description: 获取更新用户信息验证码
      operationId: update-user-info-captcha
      parameters: []
      responses:
        '200':
          description: 发送成功
          content:
            application/json:
              schema:
                type: string
      security:
        - bearer: []
      summary: ''
      tags:
        - captcha
  /user/register:
    post:
      description: 注册用户
      operationId: user-register
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterUserDto'
      responses:
        '200':
          description: 注册成功
          content:
            application/json:
              schema:
                type: string
        '400':
          description: 验证码已失效/验证码不正确/用户已存在
          content:
            application/json:
              schema:
                type: string
      summary: ''
      tags:
        - user
  /user/login:
    post:
      description: 普通用户登录
      operationId: user-login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginUserDto'
      responses:
        '200':
          description: 用户信息和token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginUserVo'
        '400':
          description: 用户不存在/密码错误
          content:
            application/json:
              schema:
                type: string
      summary: ''
      tags:
        - user
  /user/admin/login:
    post:
      description: 管理员登录
      operationId: system-login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginUserDto'
      responses:
        '200':
          description: 用户信息和token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginUserVo'
        '400':
          description: 用户不存在/密码错误
          content:
            application/json:
              schema:
                type: string
      summary: ''
      tags:
        - user
        - system
  /user/refresh:
    get:
      description: 使用refreshToken换取新token
      operationId: refresh-token
      parameters:
        - name: refreshToken
          required: true
          in: query
          description: 刷新token
          schema:
            example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
            type: string
      responses:
        '200':
          description: 刷新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Auth'
        '401':
          description: token 已失效，请重新登录
          content:
            application/json:
              schema:
                type: string
      summary: ''
      tags:
        - auth
  /user/info:
    get:
      description: 获取用户/管理员信息
      operationId: get-user-info
      parameters: []
      responses:
        '200':
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/UserDetailVo'
                  - $ref: '#/components/schemas/UserInfo'
                discriminator:
                  propertyName: type
                  mapping:
                    system: UserInfo
                    normal: UserDetailVo
          description: ''
      security:
        - bearer: []
      summary: ''
      tags:
        - user
        - system
  /user/forgot-password:
    post:
      description: 用户忘记密码
      operationId: forgot-password
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotUserPasswordDto'
      responses:
        '200':
          description: 找回密码成功
          content:
            application/json:
              schema:
                type: string
        '400':
          description: 验证码已失效/不正确
          content:
            application/json:
              schema:
                type: string
      summary: ''
      tags:
        - user
  /user/update-password:
    post:
      description: 用户/管理员更新密码
      operationId: update-password
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserPasswordDto'
      responses:
        '200':
          description: 用户/管理员更新密码
          content:
            application/json:
              schema:
                type: string
        '400':
          description: 验证码已失效/不正确
          content:
            application/json:
              schema:
                type: string
      security:
        - bearer: []
      summary: ''
      tags:
        - user
        - system
  /user/update:
    post:
      description: 用户/管理员更新用户/管理员信息
      operationId: update-user-info
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: string
        '400':
          description: 验证码不正确/已失效
          content:
            application/json:
              schema:
                type: string
      security:
        - bearer: []
      summary: ''
      tags:
        - user
        - system
  /user/freeze/{id}:
    put:
      description: 冻结用户
      operationId: freeze-user
      parameters:
        - name: id
          required: true
          in: path
          description: 冻结用户的用户ID
          schema:
            type: number
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: string
      security:
        - bearer: []
      summary: ''
      tags:
        - system
  /user/list:
    get:
      description: 用户列表
      operationId: get-user-list
      parameters:
        - name: skip
          required: true
          in: query
          description: 第几页
          schema:
            type: number
        - name: limit
          required: true
          in: query
          description: 每页多少条
          schema:
            type: number
        - name: username
          required: false
          in: query
          description: 用户名
          schema:
            type: string
        - name: nickName
          required: false
          in: query
          description: 昵称
          schema:
            type: string
        - name: email
          required: false
          in: query
          description: 邮箱地址
          schema:
            type: string
      responses:
        '200':
          description: 用户列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListVo'
      security:
        - bearer: []
      summary: ''
      tags:
        - user
  /user/upload:
    post:
      deprecated: true
      description: 上传图片
      operationId: uploadPicture
      parameters: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: string
      security:
        - bearer: []
      summary: ''
      tags:
        - file
...
info:
  title: 会议室预订系统
  description: api 接口文档
  version: '1.0'
  contact: {}
...
servers:
  - url: http://127.0.0.1:6020
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: 基于jwt的认证
  schemas:
    RegisterUserDto:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        nickName:
          type: string
        email:
          type: string
        captcha:
          type: number
      required:
        - username
        - password
        - nickName
        - email
        - captcha
    LoginUserDto:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    Permission:
      type: object
      properties:
        id:
          type: number
        code:
          type: string
          example: ccc
        description:
          type: string
          example: 访问 ccc 接口
      required:
        - id
        - code
        - description
    UserInfo:
      type: object
      properties:
        id:
          type: number
        username:
          type: string
          example: zhangsan
        nickName:
          type: string
          example: 张三
        email:
          type: string
          example: xx@xx.com
        headPic:
          type: string
          example: xxx.png
        phoneNumber:
          type: string
          example: '13233333333'
        isFrozen:
          type: boolean
        isAdmin:
          type: boolean
        createAt:
          format: date-time
          type: string
        roles:
          example:
            - 管理员
          type: array
          items:
            type: string
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/Permission'
      required:
        - id
        - username
        - nickName
        - email
        - headPic
        - phoneNumber
        - isFrozen
        - isAdmin
        - createAt
        - roles
        - permissions
    Auth:
      type: object
      properties:
        accessToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.
        refreshToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
        expiresIn:
          type: number
          example: '1800000'
      required:
        - accessToken
        - refreshToken
        - expiresIn
    LoginUserVo:
      type: object
      properties:
        userInfo:
          deprecated: true
          description: 用户信息
          allOf:
            - $ref: '#/components/schemas/UserInfo'
        auth:
          $ref: '#/components/schemas/Auth'
      required:
        - userInfo
        - auth
    UserDetailVo:
      type: object
      properties:
        id:
          type: number
        username:
          type: string
        nickName:
          type: string
        email:
          type: string
        headPic:
          type: string
        phoneNumber:
          type: string
        isFrozen:
          type: boolean
        createAt:
          format: date-time
          type: string
        isAdmin:
          type: boolean
      required:
        - id
        - username
        - email
        - phoneNumber
        - isFrozen
        - createAt
    ForgotUserPasswordDto:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        email:
          type: string
        captcha:
          type: number
      required:
        - username
        - password
        - email
        - captcha
    UpdateUserPasswordDto:
      type: object
      properties:
        password:
          type: string
        captcha:
          type: number
      required:
        - password
        - captcha
    UpdateUserDto:
      type: object
      properties:
        captcha:
          type: number
        headPic:
          type: string
        nickName:
          type: string
      required:
        - captcha
    User:
      type: object
      properties:
        id:
          type: number
        username:
          type: string
        nickName:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
        isFrozen:
          type: boolean
        headPic:
          type: string
        createAt:
          format: date-time
          type: string
      required:
        - id
        - username
        - nickName
        - email
        - phoneNumber
        - isFrozen
        - headPic
        - createAt
    UserListVo:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
        totalCount:
          type: number
      required:
        - users
        - totalCount
...